Migration of generate_default_results towards a different script
